% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesNMF.R
\name{run_bayesNMF}
\alias{run_bayesNMF}
\title{execute bayesNMF for signature deconvolution}
\usage{
run_bayesNMF(mat, tumor.type = "tumor.type", max_k = 10, n_runs = 10,
  iter_per_run = 1e+05, tol = 1e-07, a0 = 10, phi = 1, prior = "L1KL",
  hyper = FALSE)
}
\arguments{
\item{mat}{(matrix) N*M matrix of ints with N mutation frequencies for each of M samples}

\item{max_k}{(int) max hypothesized number of signatures to look for. default: 10}

\item{n_runs}{(int) number of times to run optimization code. default: 10}

\item{iter_per_run}{(int) number of iter for each run. default: 100000}

\item{tol}{(float) tolerance used to assess convergence}

\item{a0}{(int) value of hyperparameter `a`. default: 10}

\item{phi}{(int) value of hyperparameter `phi`. default: 1 (appropriate for l1-loss with exponential prior)}

\item{prior}{(chr) either L1KL (expoential priors) or L2KL (half-normal priors). default: L1KL}

\item{hyper}{(logical) if TRUE, reduce the effect of hyper-mutant samples in the signature discovery. default: FALSE}

\item{tumor_name}{(chr) name of tumor, used for plotting headers, etc. default: 'tumor.type'}
}
\value{
list of signatures, where each signature is a named list containing plots & data
}
\description{
execute bayesNMF for signature deconvolution
}

