% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesNMF.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarize results given a list of BayesNMF runs}
\usage{
summarize_results(result_list)
}
\arguments{
\item{result_list}{list of results from having called}
}
\value{
list of summarized data elements, each with length = num_runs
}
\description{
Summarize results given a list of BayesNMF runs
}

